// iti-net/src/app/models/Student.ts
export class Student{
    ID:number= 0;
    Name: string ='';
    Mobile: string = '';
    Email: string ='';
    NationalID: string='';
    Age: number=0;
}
export class StudentEditViewModel{
    ID:number= 0;
    FirstName: string ='';
    LastName: string ='';
    NameArabic: string = ''
    NameEnglish: string = ''
    Mobile: string = '';
    Email: string ='';
    NationalID: string='';
    Age: number=0;
}
export class StudentCreateViewModel{
    FirstName: string ='';
    LastName: string ='';
    Mobile: string = '';
    Email: string ='';
    NationalID: string='';
    Age: number=0;
}